+++ src/Members.js
@@ -1,0 +1,17 @@
+import React from 'react';
+export default function Members() {
+    return (
+        <div className="Members">
+            <div>
+                <div className="Member">
+                    <div className="MemberStatus offline" />
+                    Jared Hightower
+                </div>
+                <div className="Member">
+                    <div className="MemberStatus online" />
+                    cleverbot
+                </div>
+            </div>
+        </div>
+    )
+}
